Added stage exit assets
Added stage exit assets and messed with them in the places where they show up to look better. General changes to the foreground assets and prefabs, see minor changes. Be aware that this has likely derailed however we were dealing with changing those assets as the time of day changed. Sorry, but it needed to happen, and will likely keep happening. You can’t stop me.