revised final messages + tinkered w/ tatik
revised existing final messages to better reflect the current puzzles they sit next to, especially in scenes 13 and 18.
implemented + animated assets for the final messages in scenes 8 and 10 (although the scene 10 message is not visible in game yet since the person who writes it does not currently die).
revised some dialogue for the lonely climber to refer to the Witch Knight earlier in the game.
swapped the doors in the crossroads around so that the doors that lead to rooms 5 and 9 are swapped (the route to the summit is still the central door).
tinkered with the tatik arar dialogue script to bring it up to speed with the current text trigger script (although for some reason i couldn't get it working...)